Create fxmanifest.lua
This manifest file defines the FiveM resource metadata for the Lowrider module. It includes:

FX version compatibility (cerulean) and GTA5 targeting

Author and description fields for module clarity in the resource list

Declaration of the main client.lua script to be executed on the client side

Easy plug-and-play integration with other modules under your CyberRP-Scripts resource tree

This is required for the module to be recognized and correctly registered by the FiveM engine.